Akses Otoritas -> Otoritas Clear

+++ src/app/components/master/otoritas/otoritas.component.ts
@@ -3,6 +3,7 @@ import { MatDialog } from '@angular/material/dialog';
 import { PageEvent } from '@angular/material/paginator';
 import { ApiService } from 'src/app/shared/api.service';
 import { ModalOtoritasComponent } from './modal-otoritas/modal-otoritas.component';
+import { VoidComponent } from '../../modals/void/void.component';
 
 @Component({
   selector: 'app-otoritas',
@@ -10,6 +11,8 @@ import { ModalOtoritasComponent } from './modal-otoritas/modal-otoritas.componen
   styleUrls: ['./otoritas.component.css'],
 })
 export class OtoritasComponent {
+  akses = this.api.akses.role_otoritas;
+
   table = 'ms_userid/';
   dataSearch = '';
   pageSize = 50;
@@ -30,40 +33,40 @@ export class OtoritasComponent {
   }
 
   deleteData(id: number) {
-    const dialogRef = this.dialog.open(ModalOtoritasComponent, {
-      data: { name: 'delete' },
-    });
-
-    dialogRef.afterClosed().subscribe((result) => {
-      if (result === 'ya') {
-        this.api.deleteData(this.table + id).subscribe(() => {
-          this.api.getData(this.table).subscribe((res) => {
-            this.getMaxId = res[res.length - 1].id;
-          });
-          this.length = this.length - 1;
-          this.getPageData();
+    if (this.akses.edit) {
+      this.dialog
+        .open(VoidComponent)
+        .afterClosed()
+        .subscribe((result) => {
+          if (result === 'ya') {
+            this.api.deleteData(this.table + id).subscribe(() => {
+              this.length = this.length - 1;
+              this.getPageData();
+            });
+          }
         });
-      }
-    });
+    } else {
+      window.alert('Anda tidak memiliki Akses');
+    }
   }
 
   editData(data: any) {
-    const dialogRef = this.dialog.open(ModalOtoritasComponent, {
-      data: { name: 'edit', data: data },
-    });
-
-    dialogRef.afterClosed().subscribe((result) => {
-      if (result === 'simpan') {
-        this.catchResult = this.api.catchData();
-        let data = this.catchResult;
-        let id = this.catchResult.id;
-        this.api.updateData(this.table, data, id).subscribe((res) => {
-          localStorage.setItem('user', JSON.stringify(res));
-          this.getPageData();
-          window.location.reload();
-        });
-      }
-    });
+    this.dialog
+      .open(ModalOtoritasComponent, {
+        data: data,
+      })
+      .afterClosed()
+      .subscribe((result) => {
+        if (result != undefined) {
+          this.api
+            .updateData(this.table, result, result.id)
+            .subscribe((res) => {
+              localStorage.setItem('user', JSON.stringify(res));
+              this.getPageData();
+              window.location.reload();
+            });
+        }
+      });
   }
 
   handlePageEvent(event: PageEvent) {
